Renamed packets and added incoming chat packet
Renamed packets to read more naturally
