+++ mixin/1.7.10/src/main/java/pw/stamina/minecraftapi/mixin/network/outgoing/MixinChatPacket.java
@@ -22,12 +22,25 @@
  * SOFTWARE.
  */
 
-package pw.stamina.minecraftapi.network.outgoing;
+package pw.stamina.minecraftapi.mixin.network.outgoing;
 
-import pw.stamina.minecraftapi.network.Packet;
+import net.minecraft.network.play.client.C01PacketChatMessage;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import pw.stamina.minecraftapi.network.outgoing.ChatPacket;
 
-public interface PacketPlayer extends Packet {
+@Mixin(C01PacketChatMessage.class)
+public class MixinChatPacket implements ChatPacket {
 
-    boolean onGround();
-    void onGround(boolean onGround);
+    @Shadow private String message;
+
+    @Override
+    public String getMessage() {
+        return message;
+    }
+
+    @Override
+    public String setMessage(String message) {
+        return this.message = message;
+    }
 }
