+++ mixin/1.7.10/src/main/java/pw/stamina/minecraftapi/mixin/network/outgoing/MixinPositionLookPacket.java
@@ -26,9 +26,9 @@ package pw.stamina.minecraftapi.mixin.network.outgoing;
 
 import net.minecraft.network.play.client.C03PacketPlayer;
 import org.spongepowered.asm.mixin.Mixin;
-import pw.stamina.minecraftapi.network.outgoing.PacketPlayerLook;
+import pw.stamina.minecraftapi.network.outgoing.PositionLookPacket;
 
-@Mixin(C03PacketPlayer.C05PacketPlayerLook.class)
-public abstract class MixinPacketPlayerLook implements PacketPlayerLook {
+@Mixin(C03PacketPlayer.C06PacketPlayerPosLook.class)
+public abstract class MixinPositionLookPacket implements PositionLookPacket {
 
 }
