+++ mixin/1.8.9/src/main/java/pw/stamina/minecraftapi/mixin/network/outgoing/MixinChatPacket.java
@@ -24,11 +24,23 @@
 
 package pw.stamina.minecraftapi.mixin.network.outgoing;
 
-import net.minecraft.network.play.client.C03PacketPlayer;
+import net.minecraft.network.play.client.C01PacketChatMessage;
 import org.spongepowered.asm.mixin.Mixin;
-import pw.stamina.minecraftapi.network.outgoing.PacketPlayerPosition;
+import org.spongepowered.asm.mixin.Shadow;
+import pw.stamina.minecraftapi.network.outgoing.ChatPacket;
 
-@Mixin(C03PacketPlayer.C04PacketPlayerPosition.class)
-public abstract class MixinPacketPlayerPosition implements PacketPlayerPosition {
+@Mixin(C01PacketChatMessage.class)
+public class MixinChatPacket implements ChatPacket {
 
+    @Shadow private String message;
+
+    @Override
+    public String getMessage() {
+        return message;
+    }
+
+    @Override
+    public String setMessage(String message) {
+        return this.message = message;
+    }
 }
