+++ mixin/1.8.9/src/main/java/pw/stamina/minecraftapi/mixin/network/outgoing/MixinLookPacket.java
@@ -26,9 +26,9 @@ package pw.stamina.minecraftapi.mixin.network.outgoing;
 
 import net.minecraft.network.play.client.C03PacketPlayer;
 import org.spongepowered.asm.mixin.Mixin;
-import pw.stamina.minecraftapi.network.outgoing.PacketPlayerPositionLook;
+import pw.stamina.minecraftapi.network.outgoing.LookPacket;
 
-@Mixin(C03PacketPlayer.C06PacketPlayerPosLook.class)
-public abstract class MixinPacketPlayerPositionLook implements PacketPlayerPositionLook {
+@Mixin(C03PacketPlayer.C05PacketPlayerLook.class)
+public abstract class MixinLookPacket implements LookPacket {
 
 }
