+++ src/main/java/pw/stamina/minecraftapi/network/outgoing/OnGroundPacket.java
@@ -22,13 +22,12 @@
  * SOFTWARE.
  */
 
-package pw.stamina.minecraftapi.mixin.network.outgoing;
+package pw.stamina.minecraftapi.network.outgoing;
 
-import net.minecraft.network.play.client.C03PacketPlayer;
-import org.spongepowered.asm.mixin.Mixin;
-import pw.stamina.minecraftapi.network.outgoing.PacketPlayerPosition;
+import pw.stamina.minecraftapi.network.Packet;
 
-@Mixin(C03PacketPlayer.C04PacketPlayerPosition.class)
-public abstract class MixinPacketPlayerPosition implements PacketPlayerPosition {
+public interface OnGroundPacket extends Packet {
 
+    boolean onGround();
+    void onGround(boolean onGround);
 }
